删除了冗余的代码，使用共有结构
